+++ src/spicelib/devices/mos2/mos2sld.c
@@ -536,8 +536,9 @@ next1:
                     Osxpbd -= tag0 * *(here->MOS2dphibd_dl);
                     Osxpgb -= tag0 * *(here->MOS2dphigb_dl);
                 }
-                if(here->MOS2sens_w && 
-                        (iparmno == (here->MOS2senParmNo + here->MOS2sens_l))){
+                if(here->MOS2sens_w &&
+                        (iparmno == (here->MOS2senParmNo +
+                            (int) here->MOS2sens_l))){
                     Osxpgs -= tag0 * *(here->MOS2dphigs_dw);
                     Osxpgd -= tag0 * *(here->MOS2dphigd_dw);
                     Osxpbs -= tag0 * *(here->MOS2dphibs_dw);
